sharpened repo and made code for atvd4
