+++ Stefan/.gitignore
@@ -6,4 +6,4 @@ CMakeFiles
 Makefile
 Results-*/
 *.o
-
+*.txt
